+++ components/core/HeroBase.tsx
